+++ app/acciones/phpGuardarConfiguracion.php
@@ -18,16 +18,28 @@
 		$telegram_tiempo = 1;
 	}
 
-	if (!isset($_POST["notificacion_telegram"])) {
-		$notificacion_telegram = 0;
+	if (!isset($_POST["telegram_notificacion"])) {
+		$telegram_notificacion = 0;
 	}else{
-		$notificacion_telegram = 1;
+		$telegram_notificacion = 1;
 	}
 
 	if (!isset($_POST["telegram_conexion"])) {
 		$telegram_conexion = 0;
 	}else{
 		$telegram_conexion = 1;
+	}
+	
+	if (!isset($_POST["telegram_empieza_grabacion"])) {
+		$telegram_empieza_grabacion = 0;
+	}else{
+		$telegram_empieza_grabacion = 1;
+	}
+	
+	if (!isset($_POST["telegram_para_grabacion"])) {
+		$telegram_para_grabacion = 0;
+	}else{
+		$telegram_para_grabacion = 1;
 	}
 
 	$cont = 0;
@@ -56,12 +68,16 @@
 		"texto_para" => $_POST["texto_para"],
 		"texto_tiempo" => $_POST["texto_tiempo"],
 		"texto_conexion" => $_POST["texto_conexion"],
+		"texto_empieza_grabacion" => $_POST["texto_empieza_grabacion"],
+		"texto_para_grabacion" => $_POST["texto_para_grabacion"],
 		"telegram_empieza" => $telegram_empieza,
 		"telegram_para" => $telegram_para,
 		"telegram_tiempo" => $telegram_tiempo,
 		"telegram_tiempo_limite" => $_POST["telegram_tiempo_limite"],
 		"telegram_conexion" => $telegram_conexion,
-		"notificacion_telegram" => $notificacion_telegram,
+		"telegram_notificacion" => $telegram_notificacion,
+		"telegram_empieza_grabacion" => $telegram_empieza_grabacion,
+		"telegram_para_grabacion" => $telegram_para_grabacion,
 		"bot_token" => $_POST["bot_token"],
 		"id_chat" => $_POST["id_chat"],
 		"unidadTiempo" => $_POST["unidadTiempo"],